Issue #2863563 by Lendude, michielnugter, larowlan: Convert PreviewTest WebTestBase to BrowserTestBase and JavascriptTestBase
(cherry picked from commit 97ac5b70c24c80cfe1239ad73481ce68e8d007b7)
