+++ modules/views/tests/modules/views_test_config/test_views/views.view.test_pager_full_ajax.yml
@@ -1,0 +1,37 @@
+langcode: en
+status: true
+dependencies:
+  module:
+    - node
+id: test_pager_full_ajax
+label: ''
+module: views
+description: ''
+tag: ''
+base_table: node_field_data
+base_field: nid
+core: '8'
+display:
+  default:
+    display_options:
+      use_ajax: true
+      access:
+        type: none
+      cache:
+        type: tag
+      exposed_form:
+        type: basic
+      pager:
+        options:
+          id: 0
+          items_per_page: 5
+          offset: 0
+        type: full
+      style:
+        type: default
+      row:
+        type: 'entity:node'
+    display_plugin: default
+    display_title: Master
+    id: default
+    position: 0
